chore(manager): rename servers to environments
